Linkify logos and remove extraneous text:
Warren requested removal of the FreeNASÂ2018 - iXsystems, Inc text that is always displayed in the top right area of the u. It is overlapping and causing readability issue on smaller screens/resized windows.
In its place I added links to the logos that display in the upper-left area of the screen:
 - FreeNAS logo links to the FreeNAS website, iX logo links to iXsystems website, and TrueNAS logo links to the Support Overview website.