Update the Jail Wizard tooltips Updated the tooltips to the new formatting style The Jail Wizard is for simple jail configuration.
The IPv4 and IPv6 address fields only allow a single IP to be entered.
(The regex validation confirms) Updated the tooltips to reflect this.

After linking to specific sections of the User Guide works, there
may be other updates/clarifications that can be made.

Tooltips in this update are geared towards Wizard and newbie use.