Sign-in card bg image fixes
This puts the background image actually behind the logo (it was in front
before, obscuring the logo) as well as a significant reduction in file
size.  The stars bg image was originally 6000x4000px, taking up 4.8MB of
download when the displayed size is set at 400px.  This adds optional
sized images for different display resolutions, saving up to 99% (eg was
4.8MB, now 22KB).

Lastly, the accessibility of the background image is improved as well,
now having more detailed alt text.