Fix If login page appear enter "{user}" and "{password}"
Changed If login page appear enter "{user}" and "{password}" to look for //mat-list-item[@ix-auto="option__Dashboard"] first instead of login. It is to speed up the UI and to avoid looking for login when we are aloready login.

This fix the login part of test NAS-T913

(cherry picked from commit 4bee511ccfb0308797c50b10739d2cd9afb67455)