Remove unnecessary dependencies
This commit adds changes to install yarn package as that will effectively bring in latest nodejs and install npm as well. No need to define it separately. Also we are moving away from debian yarnpkg as it brings legacy nodejs dependencies which we have no use of as we have moved on to newer nodejs version.