Add files via upload
There are still lines that have been translated by me but which I don’t see in TrueNas Scale when I choose Dutch as language. The same also applies to lines/text that I still see in English in TrueNas Scale but of which I can’t find anything in the nl.json file. What happens???