NAS-125354: Bug fixes
This commit fixes unnecessary repeatative calls to fetch cloud sync credentials and cloud sync providers every time the credentials select value changes. It also fixes where the abusive file download flag checkbox was not showing up for google drive credentials because the form was only checking that for one of the Google Drive credentials where it could be many. So, turned the comparative number to a number array.