NAS-127836: Add "Netdata" button to reporting section
This button will make a WS call to generate a password and make a
request to /netdata/ for the browser to cache the credentials.
Once that is done a new tab will be opened with netdata.

(cherry picked from commit 307e3d4e85649b510ccc0a6f9754b122fff0a92a)