Fix requirements on creating certificates
TrueNAS allows creating CA certificates and CSRs. These can be used for
internal network purposes, but a CSR from TrueNAS is also
a way to get ACME-DNS authorized certificate from Let's Encrypt.

And Let's Encrypt accepts only two fields in a CSR: Common Name, and a list of SANs.
All the other subject fields are stripped from the generated certificate,
so it makes no sense to require for a certificate to have a country,
state, locality and organization.

Additionally, when creating a CA certificate, requiring SANs makes no sense,
as these are to be used for the leaf certificates.

Common Name, on the other hand, should be required for all kinds of certificates.