fix: send numeric bandwidth limit to API
the API was giving us a validation error when trying to set an rclone
bandwidth limit no matter what we put in. this was because we were
sending a string to the API instead of a number.

(cherry picked from commit 8e0f526bfc18d40fa00dc84b63f2b2351fb2c485)