NAS-138979: refactor: parallelize API calls in disk list
in the disk list, we were using `forkJoin` to combine API calls, but
`disk.details` took much longer than `disk.query` to execute, and so we
weren't seeing changes immediately after editing a single drive. after
permitting the API calls to affect the component the moment each of them
returned, we make it a little snappier.